refactor : teamRepository의 CQS패턴 적용
