+++ aiku/aiku-main/src/main/java/aiku_main/repository/TeamQueryRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TeamRepository extends JpaRepository<Team, Long>, TeamRepositoryCustom{
+public interface TeamQueryRepository extends JpaRepository<Team, Long>, TeamQueryRepositoryCustom {
+
     Optional<Team> findByIdAndStatus(Long teamId, Status status);
 }
