+++ aiku/aiku-main/src/main/java/aiku_main/repository/TeamQueryRepositoryCustom.java
@@ -1,5 +1,7 @@
 package aiku_main.repository;
 
+import aiku_main.application_event.domain.TeamResultMember;
+import aiku_main.dto.TeamEachListResDto;
 import common.domain.team.Team;
 import common.domain.team.TeamMember;
 import org.springframework.data.repository.query.Param;
@@ -7,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface TeamRepositoryCustom {
+public interface TeamQueryRepositoryCustom {
 
     Optional<Team> findTeamWithMember(Long teamId);
 
@@ -17,4 +19,7 @@ public interface TeamRepositoryCustom {
 
     boolean existTeamMember(@Param("memberId") Long memberId, @Param("teamId") Long teamId);
     Long countOfAliveTeamMember(Long teamId);
+
+    List<TeamEachListResDto> getTeamList(Long memberId, int page);
+    List<TeamResultMember> getTeamLateTimeResult(Long teamId);
 }
